+++ src/wayland/data_device/dnd_grab.rs
@@ -64,12 +64,9 @@ where
         &mut self,
         _data: &mut D,
         dh: &DisplayHandle,
-        handle: &mut PointerInnerHandle<'_, D>,
+        _handle: &mut PointerInnerHandle<'_, D>,
         event: &MotionEvent,
     ) {
-        // While the grab is active, no client has pointer focus
-        handle.motion(event.location, None, event.serial, event.time);
-
         let seat_data = self
             .seat
             .user_data()